rename annoucementName -> name

+++ src/main/java/com/FreeDoska/doska/repo/AnnouncementRepo.java
@@ -2,9 +2,10 @@ package com.FreeDoska.doska.repo;
 
 import com.FreeDoska.doska.AnnouncementModel;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface AnnouncementRepo extends JpaRepository<AnnouncementModel, Long> {
-//List<AnnouncementModel> findByTag(String tag);
+public interface AnnouncementRepo extends JpaRepository<AnnouncementModel, Integer > {
+//List<AnnouncementModel> findById(Long id);
 }
